Change up form

+++ app/nfl/forms.py
@@ -1,5 +1,5 @@
 from flask_wtf import Form 
-from wtforms import TextField, IntegerField, RadioField, HiddenField
+from wtforms import TextField, IntegerField, RadioField, HiddenField, FloatField
 from wtforms.validators import DataRequired, Length, Email, EqualTo, NumberRange
 
 class CreateNflBet(Form):
@@ -7,7 +7,7 @@ class CreateNflBet(Form):
     bet_key = HiddenField("Bet Key")
     away_team = TextField("Away Team")
     home_team = TextField("Home Team")
-    over_under = IntegerField("Over/Under", validators=[NumberRange(min=-100, max=100, message="Over/Under bet must be inbetween -100 and 100 for it to be valid.")])
+    over_under = FloatField("Over/Under", validators=[NumberRange(min=-100, max=100, message="Over/Under bet must be inbetween -100 and 100 for it to be valid.")])
     point_spread = IntegerField("Point Spread")
     home_team_ml = IntegerField("Home Team ML")
     away_team_ml = IntegerField("Away Team ML")
